# docs/swagger.yaml
basePath: /v1
definitions:
...
  dto.ResponseJson:
    properties:
      data: {}
      error:
        type: string
      limit:
        type: integer
      message:
        type: string
      offset:
        type: integer
      total_records:
        type: integer
    type: object
...
  models.Brands:
    properties:
      brand_id:
        type: string
      brand_name:
        type: string
    type: object
  models.Categories:
    properties:
      category_id:
        type: string
      category_name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: varusi0605@gmail.com
    name: varusai
  description: It's an basic ecommerce site where you can view products and order
    any product. Then also you can be a seller by listing your product and also admin
  license:
    url: https://github.com/marees7/varusai-aug-2024
  title: Shopping-site API
  version: "1.0"
paths:
  /admin/brand:
    post:
      consumes:
      - application/json
      description: Create a new brand
      operationId: create_brand
      parameters:
      - description: Enter brand details
        in: body
        name: Login
        required: true
        schema:
          $ref: '#/definitions/models.Brands'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResponseJson'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ResponseJson'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ResponseJson'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ResponseJson'
      security:
      - JWT: []
      summary: Create brand
      tags:
      - Admin
  /admin/category:
    post:
      consumes:
      - application/json
      description: Create a new category
      operationId: create_category
      parameters:
      - description: Enter category details
        in: body
        name: details
        required: true
        schema:
          $ref: '#/definitions/models.Categories'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResponseJson'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ResponseJson'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ResponseJson'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ResponseJson'
      security:
      - JWT: []
      summary: Create category
      tags:
      - Admin
...
swagger: "2.0"